Add files via upload
/*This program will read the data of an array, the array will be recognized by the letter a. the array a 
will have 20 float numbers. 
(a)It is requested by this exercise the data of this array to be printed out 
in reverse. 
(b)There will be a second array with the name b. This array will have 30 char values, it is requested to be 
counted the letter a, in this array. 
(c) at last there will be an onother array with the name c and this array will have boolean values. This 
array will have a total lenght 15. it is requested to be counted the true values and false values*/